+++ NEWS.md
@@ -1,6 +1,9 @@
 # spant 0.5.0
-* Added calc_peak_info function for simple singlet analysis.
+* A default brain PRESS basis is now simulated by the fit_mrs function when the
+basis argument isn't specified.
+* Added calc_peak_info function for simple singlet analyses.
 * crop_spec function now maintains the orignal frequency scale.
+* The basis set used for analyses has now been added to the fit result object.
 
 # spant 0.4.0
 * Bug fix for SPAR/SDAT SVS voxel dimensions.